% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrapping of Distance Matrix}
\usage{
Bootstrap(
  data,
  dist_mat_null,
  k = 10,
  kernel = c("gaussian", "euclidean"),
  normalization = c("cosine", "lognorm", "none"),
  normalize_factor = 10000,
  pca_dims = 0,
  norm_type = c("l1", "l2"),
  n_iters = 100,
  ratio = 0.05,
  t = 0,
  calc_perturb_mat = FALSE,
  n_cores = NULL,
  zero_percent = 0.7,
  ...
)
}
\arguments{
\item{data}{An M x d matrix or data.frame with M rows of data points and d columns of features.}

\item{dist_mat_null}{An M x M distance matrix calculated from the original data (null).}

\item{k}{Number of nearest neighbors. Default is 10. See details from \code{\link[RANN]{nn2}}.}

\item{kernel}{Kernel distance used:
\itemize{
  \item gaussian, gaussian distance kernel. See details \code{\link[FuseNet]{EuclideanDist}}.
  \item euclidean, euclidean distance kernel. See details \code{\link[FuseNet]{GaussianDist}}.
  }}

\item{normalization}{Normalization method used:
#' \itemize{
  \item cosine, cosine normalization. See details \code{\link[FuseNet]{Normalization}}.
  \item lognorm, log normalization. See details \code{\link[FuseNet]{Normalization}}.
  \item none, normalization is not performed.
  }}

\item{normalize_factor}{Normalize factor used in log normalization. Default is 10000. See details \code{\link[FuseNet]{Normalization}}.}

\item{pca_dims}{Number of dimensions used. Default is 0 and PCA is not performed.}

\item{norm_type}{Type of norm used:
\itemize{
  \item l1, L1-like norm. See details \code{\link[FuseNet]{L1Norm}}.
  \item l2, L1-like norm. See details \code{\link[FuseNet]{L2Norm}}.
  }}

\item{n_iters}{Number of bootstrapping iterations. Default is 100.}

\item{ratio}{Fraction of features to be downsampled in the original data matrix. Default is 0.05 aka 5\%.}

\item{t}{Matrix power used for the distance matrix. Default is 0 and powering is not performed. See \code{\link[FuseNet]{MatrixPower}} for details.}

\item{calc_perturb_mat}{Whether to calculate the perturb matrix. Default is FALSE.}

\item{n_cores}{Number of cores used. Default is to use all existing cores. See details \code{\link[parallel]{makeCluster}}.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zeros in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}

\item{...}{Additional parameters pass to \code{\link[parallel]{makeCluster}}.}
}
\value{
Returns a list with entries:
\itemize{
  \item feature_weight, n x d binary matrix with n rows of bootstrap iterations and d columns of features where 0 means feature not sampled and 1 means sampled.
  \item sample_weight, n x M matrix with n rows of bootstrap iterations and M columns of data points where each entry represents weight.
  \item perturb_mat, d x M matrix with d rows of features and M columns of data points where each entry represents the relative importance of a feature to a data point.
  \item dist_mat, M x M distance matrix.
  }
}
\description{
Bootstrapping  of Distance Matrix
}
\examples{
\dontrun{
boot.result <- Bootstrap(data, dist.mat.null, k = 10, pca_dims = 10, n_iters = 100, ratio = 0.1)
}
}
