% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{FuseData}
\alias{FuseData}
\title{Weight and Fuse Objects}
\usage{
FuseData(..., project_k = 10, zero_percent = 0.7)
}
\arguments{
\item{...}{FuseNet objects to fuse.}

\item{project_k}{Number of nearest neighbors to project based on geometric sketches, if it has been run. Default is 10.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}
}
\value{
Returns a list with entries:
\itemize{
  \item fused_weight, n x M matrix with n rows of objects or modality to fuse and M columns of data points.
  \item fused_dist, M x M fused distance matrix.
  }
}
\description{
Weight and fuse distance matrices based on the relative weights.
}
\examples{
{
object1 <- InitiateFuseNet(t(iris[,1:2]), project_name = "FuseNet", k = 3)
object1 <- RunFuseNet(object1, n_iters = 1, k = 3, ratio = 0.5, n_cores = 1)
object2 <- InitiateFuseNet(t(iris[,3:4]), project_name = "FuseNet", k = 3)
object2 <- RunFuseNet(object2, n_iters = 2, k = 3, ratio = 0.5, n_cores = 1)
fused.data <- FuseData(object1, object2)
}

}
\concept{fuse}
