% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{FuseNet-class}
\alias{FuseNet-class}
\title{The FuseNet S4 Class}
\description{
The FuseNet object with the slot information listed as follow:
}
\section{Slots}{

\describe{
\item{\code{project_name}}{Name of the project.}

\item{\code{raw_data}}{Raw data. A d x M matrix with d rows of features and M columns of data points.}

\item{\code{normalized_data}}{Normalized data. Same shape as raw_data.}

\item{\code{scaled_data}}{Scaled data (z-score). Same shape as raw_data.}

\item{\code{pca}}{Principal component analysis result, see \code{\link[irlba]{irlba}}.}

\item{\code{dist_null}}{Null nearest neighbor M x M distance matrix.}

\item{\code{sketch_id}}{Geomertric sketching cell IDs.}

\item{\code{sketch_dist}}{Geomertric sketching distance matrix.}

\item{\code{weight_mat}}{List of feature and sample weight matrices:
\itemize{
  \item feature_weight, n x d binary matrix with n rows of bootstrap iterations and d columns of features where 0 means feature not sampled and 1 means sampled.
  \item sample_weight, n x M matrix with n rows of bootstrap iterations and M columns of data points where each entry represents weight.
  \item perturb_mat, d x M matrix with d rows of features and M columns of data points where each entry represents the relative importance of a feature to a data point.
  }}

\item{\code{dist_mat}}{Permuted distance matrix.}

\item{\code{params}}{Commands used.}
}}

\concept{class}
