% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR.R
\name{FuseWLR}
\alias{FuseWLR}
\title{Weight and Fuse WLR Objects}
\usage{
FuseWLR(..., project_k = 10, zero_percent = 0.7)
}
\arguments{
\item{...}{WLR objects to fuse.}

\item{project_k}{Number of nearest neighbours to project based on geometric sketches, if it has been run. Default is 10.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}
}
\value{
Returns a list with entries:
\itemize{
  \item fused_weight, n x M matrix with n rows of objects or modality to fuse and M columns of data points.
  \item fused_dist, M x M fused distance matrix.
  }
}
\description{
Weight and fuse WLR distance matrices based on the relative weights.
}
\examples{
\donttest{
fused.data <- FuseWLR(object1, object2, project_k = 20)
}
}
