% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{InitiateFuseNet}
\alias{InitiateFuseNet}
\title{Initialize FuseNet Object}
\usage{
InitiateFuseNet(
  raw_data,
  project_name = "",
  normalization = c("cosine", "lognorm", "none"),
  normalize_factor = 10000,
  zero_percent = 0.7,
  pca_dims = 0,
  kernel = c("gaussian", "euclidean"),
  k = 100,
  t = 0,
  verbose = FALSE,
  seed = 1
)
}
\arguments{
\item{raw_data}{Raw data. An N x M matrix with N rows of features and M columns of data points.}

\item{project_name}{Project name. Default is none.}

\item{normalization}{Normalization method used. Default is cosine. See details \code{\link[FuseNet]{Normalization}}.
\itemize{
  \item cosine, cosine normalization: feature counts for each data point are divided by the L2 norm of them.
  \item lognorm, log normalization: feature counts for each data point are divided by the total sum of them. Then the data is multiplied by the scale.factor before taking a log-transformed by log(1+x).
  \item none, additional normalization is not performed.
  }}

\item{normalize_factor}{Normalization factor used with lognorm method. Default is 10000.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}

\item{pca_dims}{Number of dimensions used. Default is 0 and PCA is not performed.}

\item{kernel}{Kernel distance used:
\itemize{
  \item gaussian, gaussian distance kernel. See details \code{\link[FuseNet]{EuclideanDist}}.
  \item euclidean, euclidean distance kernel. See details \code{\link[FuseNet]{GaussianDist}}.
  }}

\item{k}{Number of nearest neighbors. Default is 100. See details from \code{\link[RANN]{nn2}}.}

\item{t}{Matrix power used for the distance matrix. Default is 0 and powering is not performed. See \code{\link[FuseNet]{MatrixPower}} for details.}

\item{verbose}{Whether to display a process bar. Default is FALSE.}

\item{seed}{Random seed number. Default is 1.}
}
\value{
Returns a FuseNet object.
}
\description{
Initialize a FuseNet object.
}
\examples{
{
object <- InitiateFuseNet(t(iris[,1:4]), project_name = "FuseNet", k = 10)
}

}
\concept{fuse}
