% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{L1Norm}
\alias{L1Norm}
\title{L1-like Norm}
\usage{
L1Norm(mat1, mat2)
}
\arguments{
\item{mat1, mat2}{Two matrices.}
}
\value{
Returns a matrix with L1-like norms.
}
\description{
Compute L1-like norm defined by the absolute values of the differences between each entry of two matrices with the same dimensions.
}
\examples{
\dontrun{
l1.norm <- L1Norm(matrix1, matrix2)
}
}
