% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{L2Norm}
\alias{L2Norm}
\title{L2-like Norm}
\usage{
L2Norm(mat1, mat2)
}
\arguments{
\item{mat1, mat2}{Two matrices.}
}
\value{
Returns a matrix with L2-like norms.
}
\description{
Compute L2-like norm defined by the square values of the differences between each entrie of two matrices with the same dimensions.
}
\examples{
\donttest{
l2.norm <- L2Norm(matrix1, matrix2)
}
}
