% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MatrixBigPower}
\alias{MatrixBigPower}
\title{Matrix Big Power}
\usage{
MatrixBigPower(A, n)
}
\arguments{
\item{A}{A matrix.}

\item{n}{Number of powers.}
}
\value{
Returns the powered matrix.
}
\description{
Compute big powers (>10) of matrix.
}
\keyword{internal}
