% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{MergeLists}
\alias{MergeLists}
\title{Merge Two List}
\usage{
MergeLists(a, b)
}
\arguments{
\item{a, b}{Two lists.}
}
\value{
Returns the merged list.
}
\description{
Merge twp lists by the name of their elements.
}
