% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Normalization}
\alias{Normalization}
\title{Data Normalization}
\usage{
Normalization(
  counts,
  normalization = c("cosine", "lognorm", "none"),
  normalize_factor = 10000,
  zero_percent = 0.7,
  verbose = FALSE
)
}
\arguments{
\item{counts}{Data to normalize. An N x M matrix with N rows of features and M columns of data points.}

\item{normalization}{Normalization method used. Default is cosine.
\itemize{
  \item cosine, cosine normalization: feature counts for each data point are divided by the L2 norm of them.
  \item lognorm, log normalization: feature counts for each data point are divided by the total sum of them. Then the data is multiplied by the scale.factor before taking a log-transformed by log(1+x).
  \item none, additional normalization is not performed.
  }}

\item{normalize_factor}{Normalization factor used with lognorm method. Default is 10000.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zeros in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}

\item{verbose}{Whether to display a process bar. Default is FALSE.}
}
\value{
Returns the normalized data.
}
\description{
Normalize count or other data.
}
\examples{
{
data.use <- matrix(runif(100), 10)
normalized.data <- Normalization(data.use)
}

}
\concept{utils}
