% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{PCA}
\alias{PCA}
\title{Principal Components Analysis}
\usage{
PCA(scaled_data, n_dims = 10, seed = 1)
}
\arguments{
\item{scaled_data}{Scaled and centered data (z-scores).}

\item{n_dims}{Number of dimensions to return. Default is 10.}

\item{seed}{Random seed number. Default is 1.}
}
\value{
Returns a list contains outputs from \code{\link[irlba]{irlba}}.
}
\description{
Perform principal components analysis on scaled and centered data (z-scores).
}
\examples{
\donttest{
pca.result <- PCA(scaled.data, n_dims = 3)
}
}
