% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ProjectSketch}
\alias{ProjectSketch}
\title{Project Geometric Sketches}
\usage{
ProjectSketch(data, data_query, weights_data, k)
}
\arguments{
\item{data}{An M x d matrix or data.frame with M rows of data points and d columns of features.}

\item{data_query}{An N x d matrix or data.frame with N rows of queried data points and d columns of features.}

\item{weights_data}{A vector of weights for each data point with total length of M.}

\item{k}{Number of nearest neighbors. See details from \code{\link[RANN]{nn2}}.}
}
\value{
Returns the weight matrix of the queried data.
}
\description{
Project weights learned from geometric sketches onto other data points based on nearest neighbors.
}
\examples{
\dontrun{
weight.mat <- ProjectSketch(data, data.query, weights.data, 30)
}
}
