% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RunGeomSketch}
\alias{RunGeomSketch}
\title{Run Geometric Sketching}
\usage{
RunGeomSketch(
  data,
  geom_size,
  is_pca = FALSE,
  n_pca = 10,
  seed = 1,
  which_python = Sys.which(names = "python3")
)
}
\arguments{
\item{data}{An M x d matrix or data.frame with M rows of data points and d columns of features.}

\item{geom_size}{Size of geometric sketches to return.}

\item{is_pca}{Whether the data columns are principal components.}

\item{n_pca}{Number of PCA dimensions to use. Default is 10.}

\item{seed}{Random seed number. Default is 1.}

\item{which_python}{Path to python3 used.}
}
\value{
Returns geometric sketch ID.
}
\description{
Run Geometric sketching sampling.
See Hie et al. (\doi{10.1016/j.cels.2019.05.003}) for details.
This function relies on reticulate R package to import geosketch python library. Only one python environment is allowed by reticulate configuration.
Therefore please be careful when importing other R packages that will invoke and set up reticulate python configuration, i.e library(Seurat), before calling GeomSketch function.
}
\examples{
\dontrun{
sketch.ids <- RunGeomSketch(data, geom_size = 1000)
}
}
