% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR.R
\name{RunWLR}
\alias{RunWLR}
\title{Run Weighted Local Robustness}
\usage{
RunWLR(
  object,
  n_iters = 100,
  ratio = 0.05,
  wlr_pca_dims = 0,
  wlr_k = 100,
  wlr_t = 0,
  norm_type = c("l1", "l2"),
  return_perturb_mat = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{object}{A WLR object.}

\item{n_iters}{Number of bootstrapping iterations. Default is 100.}

\item{ratio}{Fraction of features to be downsampled in the original data matrix. Default is 0.05 aka 5\%.}

\item{wlr_pca_dims}{Number of principle components. Default is 0 and PCA is not run.}

\item{wlr_k}{Number of nearest neighbors used. Default is 100.}

\item{wlr_t}{Matrix power used for the distance matrix. Default is 0 and powering is not performed.}

\item{norm_type}{Type of norm used:
\itemize{
  \item l1, L1-like norm. See details \code{\link[WLR]{L1Norm}}.
  \item l2, L1-like norm. See details \code{\link[WLR]{L2Norm}}.
  }}

\item{return_perturb_mat}{Whether to return the perturb matrix. Default is FALSE.}

\item{n_cores}{Number of cores used. Default is to use all existing cores. See details \code{\link[parallel]{makeCluster}}.}
}
\value{
Returns a WLR object.
}
\description{
Run Weighted Local Robustness.
}
\examples{
\donttest{
object <- RunWLR(object, n_iters = 100)
}
}
