% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{Scaling}
\alias{Scaling}
\title{Data Scaling}
\usage{
Scaling(matrix, verbose = FALSE)
}
\arguments{
\item{matrix}{Matrix to use.}

\item{verbose}{Whether to display a process bar. Default is FALSE.}
}
\value{
Returns a scaled and centered data.
}
\description{
Scale and center data.
}
\examples{
\donttest{
scaled.data <- Scaling(data)
}
}
