% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{EuclideanDist}
\alias{EuclideanDist}
\title{Euclidean Distance}
\usage{
EuclideanDist(data, ka)
}
\arguments{
\item{data}{An M x d matrix with M rows of data points and d columns of features.}

\item{ka}{Number of nearest neighbours. See details from \code{\link[RANN]{nn2}}.}
}
\value{
Returns the distance matrix.
}
\description{
Compute euclidean nearest neighbor distances.
}
\examples{
\donttest{
dist.mat <- EuclideanDist(data, ka = 10)
}
}
