% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{GaussianDist}
\alias{GaussianDist}
\title{Gaussian Distance}
\usage{
GaussianDist(data, ka)
}
\arguments{
\item{data}{An M x d matrix or data.frame with M rows of data points and d columns of features.}

\item{ka}{Number of nearest neighbours. See details from \code{\link[RANN]{nn2}}.}
}
\value{
Returns the distance matrix.
}
\description{
Compute gaussian nearest neighbor distances.
}
\examples{
\donttest{
dist.mat <- GaussianDist(data, ka = 10)
}
}
