% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR.R
\name{GeomSketch}
\alias{GeomSketch}
\title{Geometric Sketching}
\usage{
GeomSketch(
  object,
  geom_size = 1000,
  geom_pca_dims = 10,
  sketch_n_pca = 0,
  sketch_k = 30,
  sketch_t = 0,
  geom_pca_seed = 1,
  which_python = Sys.which(names = "python3")
)
}
\arguments{
\item{object}{A WLR object.}

\item{geom_size}{Size of geometric sketches to return. Default is 1000.}

\item{geom_pca_dims}{Number of PCA dimensions to use. Default is 10.}

\item{sketch_n_pca}{Number of dimensions used for the sketched data points. Default is 0 and PCA is not performed. See details from \code{\link[WLR]{InitiateWLR}}.}

\item{sketch_k}{Number of nearest neighbours for the sketched data points. Default is 30. See details from \code{\link[WLR]{InitiateWLR}}.}

\item{sketch_t}{Matrix power used for the distance matrix. Default is 0 and powering is not performed. See details from \code{\link[WLR]{InitiateWLR}}.}

\item{geom_pca_seed}{Random seed number for PCA. Default is 1.}

\item{which_python}{Path to python3 used.}
}
\value{
Returns a WLR object contains geometric sketch IDs.
}
\description{

}
\examples{
\donttest{
object <- GeomSketch(object, geom_size = 2000)
}
}
