% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLR_utils.R
\name{MatrixPower}
\alias{MatrixPower}
\title{Matrix Power}
\usage{
MatrixPower(A, n)
}
\arguments{
\item{A}{A matrix.}

\item{n}{Number of powers.}
}
\value{
Returns the powered matrix.
}
\description{
Compute powers of matrix.
}
\examples{
\donttest{
A.cubic <- MatrixPower(A, 3)
}
}
